Create TlrkLogo
Telerik Academy is considering opening a new office in Great Britain. Therefore the whole Trainers team is traveling to the United Kingdom for the important event. They’ve decided that they need to print some advertising posters and give them away to the local citizens. Please help them and print some posters with the Telerik Logo in different sizes. As a little reminder, here it is the logo itself (n. b. the letters are for reference only and denote the lengths of the sides):
Input
The input data should be read from the console. You will receive an integer number X. Note that X will always be equal to Y and Z will always be (X / 2) + 1. (Refer to the examples).
Output
The output should be printed on the console.
Use the “*” (asterisk) character to print the logo and “.” (dot) for the rest.
Constraints
•	X will always be a positive odd number between 3 and 27 inclusive.
•	Allowed working time for your program: 0.1 seconds.
•	Allowed memory: 16 MB.
Examples
Example
input	
3

Example
output
	
.*...*.
*.*.*.*
...*...
..*.*..
.*...*.
..*.*..
...*...
Example input	
5	

Example output

..*.......*..
.*.*.....*.*.
*...*...*...*
.....*.*.....
......*......
.....*.*.....
....*...*....
...*.....*...
..*.......*..
...*.....*...
....*...*....
.....*.*.....
......*......